// management/registry-ui/src/generated/client/models/model_Plan.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type model_Plan = {
    created_at?: string;
    device_limit?: number;
    duration_days?: number;
    enabled?: boolean;
    id?: number;
    name?: string;
    price?: number;
    sort_order?: number;
    speed_limit?: number;
    traffic_limit?: number;
    updated_at?: string;
};